Attempted to fix tests. Selenium still broken, MyRestService tests in Client partially working

+++ MPR_Project_Client/src/main/java/pl/edu/pjatk/MPR_Project/service/MyRestService.java
@@ -1,6 +1,7 @@
 package pl.edu.pjatk.MPR_Project.service;
 
 import jakarta.servlet.http.HttpServletResponse;
+import org.apache.coyote.Response;
 import org.apache.pdfbox.Loader;
 import org.apache.pdfbox.io.RandomAccessReadBuffer;
 import org.apache.pdfbox.pdmodel.PDDocument;
@@ -11,21 +12,24 @@ import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClient;
+import pl.edu.pjatk.MPR_Project.exception.CapybaraAlreadyExists;
+import pl.edu.pjatk.MPR_Project.exception.CapybaraNotFoundException;
 import pl.edu.pjatk.MPR_Project.exception.InvalidInputCapybaraException;
 import pl.edu.pjatk.MPR_Project.model.Capybara;
 
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Service
 public class MyRestService {
     private final StringService stringService;
-    RestClient restClient;
     private static final Logger logger = LoggerFactory.getLogger(MyRestService.class);
 
-
     @Autowired
+    RestClient restClient;
+
     public MyRestService(StringService stringService, RestClient restClient) {
         this.stringService = stringService;
         this.restClient = restClient;
@@ -33,6 +37,17 @@ public class MyRestService {
 
     public void addCapybara(Capybara capybara) {
         logger.info("Attempting to add capybara with values: {}", capybara);
+
+        Optional<Capybara> existingCapybara = restClient.get()
+                .uri("/capybara/find/identification/" + capybara.getIdentification())
+                .retrieve()
+                .body(new ParameterizedTypeReference<>() {
+                });
+
+        if (existingCapybara.isPresent()) {
+            throw new CapybaraAlreadyExists();
+        }
+
         try {
             restClient.post()
                     .uri("/capybara/add")
@@ -49,6 +64,16 @@ public class MyRestService {
 
     public void patchCapybaraById(Capybara capybara, Long id) {
         logger.info("Attempting to update capybara with ID: {} with values: {}", id, capybara);
+
+        Optional<Capybara> existingCapybara = restClient.get()
+                .uri("/capybara/find/identification/" + capybara.getIdentification())
+                .retrieve()
+                .body(new ParameterizedTypeReference<>() {
+                });
+
+        if (existingCapybara.isPresent()) {
+            throw new CapybaraAlreadyExists();
+        }
 
         if (capybara.getName().isBlank() || capybara.getAge() <= 0) {
             logger.error("Invalid Input. Failed.");
@@ -68,6 +93,17 @@ public class MyRestService {
 
     public void deleteCapybaraById(Long id) {
         logger.info("Attempting to delete capybara with ID: {}", id);
+
+        ResponseEntity<Capybara> response = restClient.get()
+                .uri("/capybara/find/id/{id}", id)
+                .retrieve()
+                .toEntity(Capybara.class);
+
+        if(response.getStatusCode() == HttpStatus.NOT_FOUND) {
+            logger.error("Failed to delete capybara. Capybara not found");
+            throw new CapybaraNotFoundException();
+        }
+
         try {
             restClient.delete()
                     .uri("/capybara/delete/{id}", id)
@@ -82,12 +118,22 @@ public class MyRestService {
 
     public List<Capybara> getByName(String name) {
         logger.info("Attempting to get capybara with name: {}", name);
+
+        ResponseEntity<Capybara> response = restClient.get()
+                .uri("/capybara/find/name/{name}", name)
+                .retrieve()
+                .toEntity(Capybara.class);
+
+        if(response.getStatusCode() == HttpStatus.NOT_FOUND) {
+            logger.error("Failed to get capybara by name. Capybara not found");
+            throw new CapybaraNotFoundException();
+        }
+
         try {
             List<Capybara> capybaraList = restClient.get()
                     .uri("/capybara/find/name/{name}", name)
                     .retrieve()
-                    .body(new ParameterizedTypeReference<>() {
-                    });
+                    .body(new ParameterizedTypeReference<>() {});
             logger.info("Successfully got capybara with name: {}", name);
             return capybaraList;
         } catch (Exception e) {
